Refactor to idiomatic imports and move environment and stackframe to bytecode crate
